break out consensus NetworkConfig into its own file
